requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.12
mangum==0.19.0
fonttools==4.61.1
fpdf2==2.8.7
google-auth==2.48.0
google-genai==1.66.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
langchain-text-splitters==0.3.8
numpy==2.2.4
openai==1.82.0
pillow==12.1.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.25.5
pymupdf4llm==0.0.17
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
sentence-transformers==4.1.0
sniffio==1.3.1
supabase==2.15.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.2
websockets==16.0
